Add files via upload
This is the Java class that implements the Cryptographic functions of the system for the the Android application. It has been broken out separately so it is easy to find the member functions of the class. Included are the functions for Asymmetric encryption and decoding, Symmetric encryption and decoding, Hashing, key generation, Digital Signature creation and Digital Signature verification.